# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: zTech CRM API
  description: API del CRM de salones de eventos corporativos — UNLaM GADS1.
  version: v1
servers:
- url: http://localhost:53103
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: Obtiene un usuario del tenant (sólo ADMIN)
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Usuario encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Rol insuficiente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Usuario inexistente o de otro tenant
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
      - Users
      summary: "Actualiza perfil, rol y estado (sólo ADMIN)"
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: Lista usuarios del tenant (sólo ADMIN)
      operationId: list
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: role
        in: query
        required: false
        schema:
          type: string
          enum:
          - ADMIN
          - SELLER
          - SALES_MANAGER
      - name: active
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "lastName,asc"
      responses:
        "200":
          description: Listado paginado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseUserResponse"
        "403":
          description: Rol insuficiente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseUserResponse"
    post:
      tags:
      - Users
      summary: Crea un usuario con contraseña temporal (sólo ADMIN)
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/users/{id}/reset-password:
    post:
      tags:
      - Users
      summary: Restablece la contraseña temporal e invalida sesiones (sólo ADMIN)
      operationId: resetPassword
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
...
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      summary: Inicia sesión y devuelve un JWT
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Login exitoso
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Datos de la solicitud inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciales inválidas
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
      security: []
  /api/v1/auth/change-password:
    post:
      tags:
      - Auth
      summary: Cambia la contraseña del usuario autenticado e invalida sus sesiones
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "204":
          description: Contraseña actualizada
        "422":
          description: La contraseña actual es incorrecta o no cambió
...
  /api/v1/opportunities/{id}/activities:
    get:
      tags:
      - Activities
      summary: Lista actividades de una oportunidad visible
      operationId: byOpportunity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "occurredAt,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseActivityResponse"
...
  /api/v1/contacts/{id}/activities:
    get:
      tags:
      - Activities
      summary: Lista actividades de un contacto visible
      operationId: byContact
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "occurredAt,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseActivityResponse"
  /api/v1/companies/{id}/activities:
    get:
      tags:
      - Activities
      summary: Lista actividades de una empresa visible
      operationId: byCompany
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "occurredAt,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseActivityResponse"
...
  /api/v1/activities:
    get:
      tags:
      - Activities
      summary: "Lista actividades visibles, ordenadas y paginadas"
      operationId: list_11
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "occurredAt,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseActivityResponse"
components:
  schemas:
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          type: string
          enum:
          - ADMIN
          - SELLER
          - SALES_MANAGER
        active:
          type: boolean
      required:
      - email
      - firstName
      - lastName
      - role
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - SELLER
          - SALES_MANAGER
        active:
          type: boolean
        mustChangePassword:
          type: boolean
...
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          type: string
          enum:
          - ADMIN
          - SELLER
          - SALES_MANAGER
        temporaryPassword:
          type: string
      required:
      - email
      - firstName
      - lastName
      - role
    ResetPasswordRequest:
      type: object
      properties:
        temporaryPassword:
          type: string
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Email del usuario
          example: admin@ztech.local
          minLength: 1
        password:
          type: string
          description: "Contraseña en texto plano, sólo para este request"
          example: Admin123!
          minLength: 1
      required:
      - email
      - password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "JWT a enviar como 'Authorization: Bearer <token>' en el resto\
            \ de los requests"
        tokenType:
          type: string
          description: Tipo de token
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int64
          description: Segundos hasta que el token expira
          example: 28800
        mustChangePassword:
          type: boolean
          description: Indica que sólo se permite cambiar la contraseña antes de continuar
        user:
          $ref: "#/components/schemas/UserSummary"
          description: "Datos mínimos del usuario autenticado, para que el frontend\
            \ no tenga que decodificar el JWT"
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
      required:
      - currentPassword
...
    PageResponseUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    UserSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
    ActivityResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/ActivityTypeResponse"
        occurredAt:
          type: string
          format: date-time
        companyId:
          type: integer
          format: int64
        contactId:
          type: integer
          format: int64
        opportunityId:
          type: integer
          format: int64
        description:
          type: string
        result:
          type: string
        author:
          $ref: "#/components/schemas/UserSummaryResponse"
        createdAt:
          type: string
          format: date-time
    ActivityTypeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        active:
          type: boolean
    PageResponseActivityResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ActivityResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT